WEB - 렌트 페이지 상세보기 & 검색 & 수정 & 삭제

+++ tracky-core/src/main/java/kernel360/trackycore/core/common/entity/CarEntity.java
@@ -94,7 +94,7 @@ public class CarEntity extends DateBaseEntity {
 		this.sum = sum;
 	}
 
-	public void updateSum(double updateSum) {
+	public void updateSum(int updateSum) {
 		this.sum += updateSum;
 	}
 }
